books and user reading intervals

+++ database/migrations/2024_03_29_170133_create_books_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name')->unique();
             $table->string('author');
-            $table->integer('number_of_pages');
+            $table->unsignedSmallInteger('number_of_pages');
             $table->timestamps();
         });
     }
